Se cambia el tipo de los parámetros de todos los métodos, desde int a double

+++ src/main/java/Calculadora.java
@@ -1,25 +1,25 @@
 public class Calculadora {
-    public static int sumar(int numero1, int numero2) {
+    public static double sumar(double numero1, double numero2) {
         return numero1 + numero2;
     }
 
-    public static int restar(int numero1, int numero2) {
+    public static double restar(double numero1, double numero2) {
         return numero1 - numero2;
     }
 
-    public static int multiplicar(int numero1, int numero2) {
+    public static double multiplicar(double numero1, double numero2) {
         return numero1 * numero2;
     }
 
     public static double dividir(double numero1, double numero2) {
-        return (float) numero1 / numero2;
+        return numero1 / numero2;
     }
 
-    public static int mayor(int numero1, int numero2) {
+    public static double mayor(double numero1, double numero2) {
         return Math.max(numero1, numero2);
     }
 
-    public static int menor(int numero1, int numero2) {
+    public static double menor(double numero1, double numero2) {
         return Math.min(numero1, numero2);
     }
 
@@ -30,5 +30,4 @@ public class Calculadora {
     public static double porcentaje(double numero, double porcentaje) {
         return (porcentaje * numero) / 100;
     }
-
 }
